improve types and fix circular dependencies
